correcciones estructurales en componentes modulo divisas
